Yogafitness Responsive HTML, CSS, and Bootstrap website
Some of the key features and assets of Eugalife are a full screen slider with parallax effect, sticky navigation, video support, pricing tables, and a clean and full-width gallery. Eugalife also has a contact page with the best class pages, a testimonial slider, an events section, and a working form. Make the most of your yoga career and grow your yoga business.